Refactored and changed joinContigs return value

+++ KmerMapper.hpp
@@ -75,7 +75,7 @@ typedef google::sparse_hash_map<Kmer, ContigRef, KmerHash> hmap_contig_t;
   void printContigs();
   int writeContigs(int count1, string contigfilename, string graphfilename);
 
-  int joinContigs(ContigRef a, ContigRef b, int a_direction, int b_direction);
+  size_t joinContigs(ContigRef a, ContigRef b, int a_direction, int b_direction);
   size_t joinContigs();
   pair<size_t, size_t> splitContigs(); 
   pair<pair<size_t, size_t>, size_t> splitAndJoinContigs();
